pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zerotrust-ai"
version = "0.1.0"
description = "AI-Accelerated Zero Trust Architecture Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Corey A. Wade", email = "corey@coreywade.com"},
]
keywords = ["zero-trust", "cybersecurity", "AI", "behavioral-analytics", "microsegmentation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking",
]
dependencies = [
    "numpy>=1.24.0",
    "flask>=3.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
zerotrust-ai = "zerotrust_ai.cli:main"

[project.urls]
Homepage = "https://github.com/cwccie/zerotrust-ai"
Repository = "https://github.com/cwccie/zerotrust-ai"
Issues = "https://github.com/cwccie/zerotrust-ai/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
